added api2 stream/next endpoint
