+++ roundware/api1/views.py
@@ -73,7 +73,7 @@ def operation_to_function(operation):
         "get_available_assets": commands.get_available_assets,
         "play_asset_in_stream": commands.play_asset_in_stream,
         "vote_asset": commands.vote_asset,
-        "skip_ahead": commands.skip_ahead,
+        "skip_ahead": api.skip_ahead,
         "get_events": commands.get_events,
     }
     key = string.lower(operation)
